Time: 2 ms (54.47%) | Memory: 10.2 MB (21.06%) - LeetSync

+++ 2076-sum-of-digits-of-string-after-convert/sum-of-digits-of-string-after-convert.cpp
@@ -8,7 +8,7 @@ public:
         }
  
         while(k>0){
-            long long num = 0;            
+            int num = 0;            
             for(char c: str){
                 num += (c - '0');
             }
@@ -17,7 +17,7 @@ public:
             k--;
         }
 
-        return stoll(str);
+        return stoi(str);
     }
 };
